Fix warnings in libziparchive
system/core/include is included in the global include path using
-isystem, which hides all warnings.  zlib.h is included through
system/core/include/ziparchive/zip_archive.h, which was hiding warnings
in it.  Use a #pragma around the call to deflateInit2, it is a macro
that expands to an old-style cast, in preparation for moving from
-isystem to -I.  Also fix an implicit conversion from size_t to
uint16_t.

Test: m -j native
Bug: 31492149

+++ include/ziparchive/zip_archive.h
@@ -44,7 +44,8 @@ struct ZipString {
    * entry_name has to be an c-style string with only ASCII characters.
    */
   explicit ZipString(const char* entry_name)
-      : name(reinterpret_cast<const uint8_t*>(entry_name)), name_length(strlen(entry_name)) {}
+      : name(reinterpret_cast<const uint8_t*>(entry_name)),
+      name_length(static_cast<uint16_t>(strlen(entry_name))) {}
 
   bool operator==(const ZipString& rhs) const {
     return name && (name_length == rhs.name_length) &&
